# openapi/auth-service.yaml
openapi: 3.1.0
info:
  title: Auth Service API
  description: "Регистрация, вход, JWT, OAuth 2.1, JWKS"
  version: "1.0"
servers:
- url: http://localhost:8086
  description: Generated server url
paths:
  /auth/register:
    post:
      tags:
      - auth-controller
      operationId: register
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
        required: true
      responses:
        "201":
          description: Created
  /auth/refresh:
    post:
      tags:
      - auth-controller
      operationId: refresh
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TokenResponse"
  /auth/login:
    post:
      tags:
      - auth-controller
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TokenResponse"
  /auth/confirm:
    get:
      tags:
      - auth-controller
      operationId: confirmByLink
      parameters:
      - name: token
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
    post:
      tags:
      - auth-controller
      operationId: confirm
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfirmRequest"
        required: true
      responses:
        "200":
          description: OK
  /.well-known/jwks.json:
    get:
      tags:
      - jwks-controller
      operationId: jwks
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
components:
  schemas:
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
          maxLength: 255
          minLength: 0
        username:
          type: string
          maxLength: 100
          minLength: 3
        password:
          type: string
          maxLength: 72
          minLength: 8
      required:
      - email
      - password
      - username
    RefreshRequest:
      type: object
      properties:
        refreshToken:
          type: string
      required:
      - refreshToken
    TokenResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        tokenType:
          type: string
        expiresIn:
          type: integer
          format: int64
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
      - email
      - password
    ConfirmRequest:
      type: object
      properties:
        token:
          type: string
      required:
      - token
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT