# services/auth-service/src/main/java/ru/potekhincode/auth/controller/AuthController.java
package ru.potekhincode.auth.controller;

import jakarta.validation.Valid;
import org.springframework.http.HttpStatus;
import org.springframework.web.bind.annotation.*;
import ru.potekhincode.auth.dto.*;
import ru.potekhincode.auth.service.AuthService;

@RestController
@RequestMapping("/auth")
public class AuthController {

    private final AuthService authService;

    public AuthController(AuthService authService) {
        this.authService = authService;
    }

    @PostMapping("/register")
    @ResponseStatus(HttpStatus.CREATED)
    public void register(@Valid @RequestBody RegisterRequest request) {
        authService.register(request);
    }

    @PostMapping("/login")
    public TokenResponse login(@Valid @RequestBody LoginRequest request) {
        return authService.login(request);
    }

    @PostMapping("/refresh")
    public TokenResponse refresh(@Valid @RequestBody RefreshRequest request) {
        return authService.refresh(request);
    }

    @PostMapping("/confirm")
    public void confirm(@Valid @RequestBody ConfirmRequest request) {
        authService.confirm(request);
    }

    @GetMapping("/confirm")
    public String confirmByLink(@RequestParam String token) {
        authService.confirm(new ConfirmRequest(token));
        return "Email подтверждён. Теперь можно войти.";
    }
}